Fix: break require cycles
