+++ src/component/geometry/ColdGeometry.tsx
@@ -11,6 +11,7 @@ import type { Feature, LineString, Polygon } from 'geojson';
 import { StoreContext } from '../../state/StoreContext';
 import { StyleContext } from '../StyleContext';
 import { orderShapes } from '../../util/geometry/display';
+import { NONPOINT_ZINDEX_PROPERTY } from '../../util/interaction';
 import type {
   RenderNonPointFeatureCollection,
   RenderProperties,
@@ -23,11 +24,6 @@ import type { Comparator } from '../../util/collections';
 export type ShapeComparator = Comparator<
   Feature<Polygon | LineString, RenderProperties>
 >;
-
-/**
- * The GeoJSON property giving the height index for rendering
- */
-export const COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY = 'rnmgeZIndex';
 
 /**
  * The ID of the bottommost points layer is referred to by other layers.
@@ -103,7 +99,7 @@ function NonPointLayers({
     list.forEach((feature) => {
       const indexFromEnd = arr.length - index - 1;
       const height = COLD_NON_POINT_LAYER_PAIR_COUNT - indexFromEnd - 1;
-      feature.properties[COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY] = height;
+      feature.properties[NONPOINT_ZINDEX_PROPERTY] = height;
     });
   });
 
@@ -127,7 +123,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['<=', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 0],
+          ['<=', ['get', NONPOINT_ZINDEX_PROPERTY], 0],
         ]}
         style={fillLayerStyle}
       />
@@ -137,7 +133,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['<=', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 0],
+          ['<=', ['get', NONPOINT_ZINDEX_PROPERTY], 0],
         ]}
         style={lineLayerStyle}
       />
@@ -147,7 +143,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 1],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 1],
         ]}
         style={fillLayerStyle}
       />
@@ -157,7 +153,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 1],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 1],
         ]}
         style={lineLayerStyle}
       />
@@ -167,7 +163,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 2],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 2],
         ]}
         style={fillLayerStyle}
       />
@@ -177,7 +173,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 2],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 2],
         ]}
         style={lineLayerStyle}
       />
@@ -187,7 +183,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 3],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 3],
         ]}
         style={fillLayerStyle}
       />
@@ -197,7 +193,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 3],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 3],
         ]}
         style={lineLayerStyle}
       />
@@ -207,7 +203,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 4],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 4],
         ]}
         style={fillLayerStyle}
       />
@@ -217,7 +213,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 4],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 4],
         ]}
         style={lineLayerStyle}
       />
@@ -227,7 +223,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 5],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 5],
         ]}
         style={fillLayerStyle}
       />
@@ -237,7 +233,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 5],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 5],
         ]}
         style={lineLayerStyle}
       />
@@ -247,7 +243,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 6],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 6],
         ]}
         style={fillLayerStyle}
       />
@@ -257,7 +253,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 6],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 6],
         ]}
         style={lineLayerStyle}
       />
@@ -267,7 +263,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 7],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 7],
         ]}
         style={fillLayerStyle}
       />
@@ -277,7 +273,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 7],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 7],
         ]}
         style={lineLayerStyle}
       />
@@ -287,7 +283,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 8],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 8],
         ]}
         style={fillLayerStyle}
       />
@@ -297,7 +293,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 8],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 8],
         ]}
         style={lineLayerStyle}
       />
@@ -307,7 +303,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'Polygon'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 9],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 9],
         ]}
         style={fillLayerStyle}
       />
@@ -317,7 +313,7 @@ function NonPointLayers({
         filter={[
           'all',
           ['==', ['geometry-type'], 'LineString'],
-          ['==', ['get', COLD_GEOMETRY_NONPOINT_ZINDEX_PROPERTY], 9],
+          ['==', ['get', NONPOINT_ZINDEX_PROPERTY], 9],
         ]}
         style={lineLayerStyle}
       />
